Adjust argument type for SDT function call when result is already a like type
Summary:
In some instances we infer `~t` for a generic parameter instead of the better `t`, which causes errors in downstream code that has to satisfy a `where` constraint on a method.

Here's an example:
```
<<__SupportDynamicType>>
function expect<T as supportdyn<mixed> >(T $obj)[]: ~ExpectObj<T> {
  throw new Exception("A");
}

<<__SupportDynamicType>>
final class ExpectObj<T as supportdyn<mixed> > {
  final public function toBeGreaterThan(num $expected): void where T as num {
  }
}

function getVec():~vec<int> { return vec[]; }
function testit():void {
  $y = getVec()[0];
  // This works
  $z = expect<int>($y);
  $z->toBeGreaterThan(0);
  // This does not, because we infer ~int for the type argument
  // This is due to the call to expect generating ~int <: #0 <: supportdyn<mixed>
  $w = expect($y);
  $w->toBeGreaterThan(0);
}
```
So the first call to `expect` followed by `toBeGreaterThan` is accepted, with explicit type argument `int`. But the second call will use `SDT` function application on `expect` that succeeds with the fully static call under `T:=~int`. Unfortunately, it's actually better to infer `int` but this relies on "distributing" the type `~int` of the argument across the static type of `expect<int>` (i.e. `int -> ~ExpectObj<int>`) and the dynamic type that comes with the `SDT` annotation (i.e. `dynamic -> dynamic`).

Observe that if `expect` had the stronger type `function(~T):~ExpectObj<T>` then all would be fine, as we would infer `int` for `T`. In the situation that the return type is *already* a like type, it's safe to transform `supportdyn<(function(t):~u)>` to `supportdyn<(function(~t):~u)>` before checking the call.

This is not entirely satisfactory, but does solve a common inference issue.

Differential Revision: D37347111

fbshipit-source-id: de3d813775b164dd15638c255aef755401c7a45e