Various opcodes
Summary:
This Diff includes addition of new opcodes and also some re-organization. Before I used methods like `assemble_local_carrying_opcode` -- as in I organized assembling just based on an opcode's parameters. I'm now moving to methods such as `assemble_cget` where the assembling is organized sort of by families of instructions. Even if a family of instructions have a lot of variation in what the variants carry, I think that this code is more readable and more understandable for anyone in the future who comes from more familiarity with `HHAS`.
Added:
- `Throw`
- `InstanceOfD`
- `CreateCl`
- `Resolve` family
- `NewVec`/`NewDict` etc
- `Set` instrs
- `Verify` instrs
- `CGet` family

Can now assemble 402/904 of the files in quick.

Reviewed By: aorenste

Differential Revision: D37559740

fbshipit-source-id: c61a47f212204dfd30034f341ebb4a050a838868