Adding class methods
Summary:
Added support for assembling methods, which are almost identical to functions but not completely (different flags and fields, like `Visibility`). Can now assemble snippets like:
```
.method {}{} [public static] (12,14) <"" N > f() {
    .srcloc 13:10,13:16;
    String "Foo\n"
    .srcloc 13:5,13:17;
    Print
    .srcloc 13:5,13:17;
    PopC
    .srcloc 14:4,14:4;
    Null
    RetC
  }
```
Also added a convenience function `expect_identfier_into_ffi_str`. A goal for the future is to increase code cleanup -- in the next Diff more on this,
Can now assemble 297/904 of the files in quick.

Reviewed By: aorenste

Differential Revision: D37555057

fbshipit-source-id: 9e98602bb79e6efd3f754b7aefb6fb59daa43443