Define a __Docs attribute on classes
Summary:
Define a `<<__Docs("http://example.com/my_project")>>` attribute that allows users to mark the docs homepage for their classes. This makes docs URLs visible to the toolchain, which will enable them to be inherited and shown in the IDE (see later diffs in this stack).

This diff introduces the `__Docs` attribute, adds the field to the AAST and updates the lowerer to set this field.

Ignore hhvm-and-hack-ubuntu which is failing on trunk too.

Reviewed By: hgoldstein

Differential Revision: D37003893

fbshipit-source-id: 131c4174f4d11a36c4f82206a92ad0b9f232d560