Rework FmtCommaSep -> FmtSep
Summary:
Rework `FmtCommaSep` into a more generic `FmtSep` with a `comma` helper that does what `FmtCommaSep` used to do.

FmtCommaSep allowed having a prefix and postfix but always separated by commas - but a minor change allows it to separate by any infix string and be more general.

Also fixed a bug in FmtIdentifier::display() which incorrectly decided if the identifier needed quotes or not.

hhvm-ubuntu is just failing for everything.

Differential Revision: D37525245

fbshipit-source-id: 750d93291a163653f37410bc8114f650d3efa570