Allow for returning values from fini() in extern-worker
Summary:
Add the ability for the fini method in an extern-worker job to return
a value. This is useful for the case when a job produces some data in
aggregate that will always be processed at once. There's less overhead
for it to be returned in a single Ref, rather than N Refs for every
input.

Use this new ability in package.cpp, along with the ability to specify
aggregate inputs with init method, for the FileMetas and
ParseMetas. These are always processed together, so there's no need
for every meta to get its own Ref. In fact, its detrimental to
performance.

Reviewed By: edwinsmith

Differential Revision: D37677180

fbshipit-source-id: 3858bd9fa3f6c4d72b8d1792ffb5dc3ff2c0bd7e