move verify_select out of main match
Summary:
Code cleanup: The verify step for Special::Select has gotten long so move it to its own function.

In addition the code for checking the prior Instr was unnecessarily complex - it really just needs to know that the prior Instr actually expected a Select - we already check the count when we check the prior Instr itself.

Differential Revision: D37594336

fbshipit-source-id: 96f46d69c72b454c9d4661570a172819be615a06