Fix indices used to get reified classes in fb_setprofile callback
Summary: In `getReifiedClasses` (which is called when the func/meth has at least one reified generic), the ArrayData pointed to by `auto const generics = val(loc).parr;` contains all the type parameters (both reified and not) for the func/method call. This means we should be incrementing the index in both cases (i.e. corresponding to iterating over each `pinfo` in `tparams.m_typeParamInfo`) so we can properly index into `generics` when needed.

Reviewed By: ricklavoie

Differential Revision: D37693759

fbshipit-source-id: c4629e1c3a5efdc4b0a4c20772d178e7818b1a80