Move m_repo out of RepoOptionsFlags and into RepoOptions
Summary:
RepoOptionsFlags is meant solely for things which are used by
HackC during parsing. m_repo is not, so it should be in RepoOptions
instead. We only need m_repo when creating a HhvmDeclProvider.

Instead of retrieving it from RepoOptionsFlags and when we don't have
access to the original RepoOptions, instead store it in
LazyUnitContentsLoader and retrieve it from there.

Reviewed By: edwinsmith

Differential Revision: D37699395

fbshipit-source-id: c6e3d2764554d4e148e87dd173d627292202d1ac