StructDicts with unaligned typed-values
Summary: This diff introduces a new memory layout for `StructDict`s. In x86 machines, we store `StructDict`s as foloows: `StructDict` header, followed by `numFields` `UnalignedTypedValue`s, followed by `numFields` iterator positions. Apart from being more cache friendly, this layout allows us to increase the size of the `StructDict` as we do not need to store the offset to values in one byte in the header.

Reviewed By: mofarrell

Differential Revision: D35802503

fbshipit-source-id: 583cb938fd9dacee0465af5cf9f93396549e6e12