Records kind of analysis result
Summary:
We need to emit the kind of shape result so that the codemod can take the appropriate action. For example, if it is an allocation, we only need to change `dict[...]` into `shape(...)`, but in case of a parameter, we need to replace the `dict` hint with a `shape` one.

This diff threads the kind information from the marker constraints into results.

Since `Argument` markers are used for testing and debugging purposes for the moment, this diff also makes the codemod directive generated ignore these results.

Reviewed By: enetsee

Differential Revision: D37614145

fbshipit-source-id: d2aae74d82ee177ad070dfddfdb99356b9af5956