Decl_inherit + Typing_extends: undo decl change, consider synthesized members when building inheritance map
Summary:
- Undoes the inheritance change (still gated + unreleased) from D37464394 (https://github.com/facebook/hhvm/commit/e23daa9986f3db9b2b4d153f6fc3c0cdade63dc8)

That diff made it so that, when `enable_strict_const_semantics = 3` is set, synthesized members do not override non-synthesized members (i.e. a require extends does not beat extends), but this is not always desirable. An example is the added test case
```
interface I {
  abstract const type T as arraykey;
}

class C implements I {
  const type T = int;
}

trait T implements I {
  require extends C;

  public function foo(): this::T {
    return 3;
  }
}
```
In `T`, if we do not allow `C::T (synthesized)` to win over `I::T`, we get a type error at `return 3;`.

It's possible that we should allow synthesized overrides in traits and interfaces but not classes, or that we should prevent synthesized overrides only in the case when the origin is the same, but this deserves its own consideration and is out of scope for this diff. Created tracking task T125402906.
 ---

- Removes early filtering of synthesized type constants when building the type constants map

This causes us to check overrides on inherited synthesized members too. It also results in a slightly better developer experience where you get stronger information from `require`. Rough example:

```
class C {
  const type T = int;
}
interface I {
  abstract const type T as arraykey;
}
trait T implements I {
  require extends C;

  // previously required in some hierarchies to refine arraykey to int
  // now an error that we are overriding concrete C::T with abstract
  abstract const type T as int;
}
class D extends C { use T; }
```
This isn't strictly an error since the abstract declaration loses to the concrete one, but it lets the inheritance checks proceed as though trait `T` has `T = int` from C. I've moved the filtering of synthesized members down to the conflict check where it matters (only real declarations can cause a conflict).

Differential Revision: D37656012

fbshipit-source-id: cd492fccc666c2db4cb7c89b8f1c226b259af3b7