allow a bigger low arena in non-lowptr builds
Summary:
In non-lowptr builds, low_malloc() is allowed to return addresses that don't fit in 32bit. So there is no point limiting the low arena to below 4G and fallback to high arena (which can cause issues in low_free). Here we enlarge the low arena in non-lowptr builds and get rid of the fallback.
The primary benefit is that we won't need to use emergency arena when the <4G address space is exhausted (the real benefit of nonlow-ptr build).

Reviewed By: ottoni

Differential Revision: D37790304

fbshipit-source-id: 453db31673c21bf11a393727ad325050997d723e