changed cmp_unit to accept an aboslute path that represents a relative path. 875/904
Summary:
A go-around solution to verify errors such as:
```
2 (/home/almathaler/fbsource/fbcode/hphp/test/quick/redeclared_class_a.php): units mismatch:
unit.symbol_refs.includes["0"]: Mismatch SearchPathRelative(Slice[114, 101, 100, 101, 99, 108, 97, 114, 101, 100, 95, 99, 108, 97, 115, 115, 46, 49, 46, 105, 110, 99])
vs
Absolute(Slice[47, 104, 111, 109, 101, 47, 97, 108, 109, 97, 116, 104, 97, 108, 101, 114, 47, 102, 98, 115, 111, 117, 114, 99, 101, 47, 102, 98, 99, 111, 100, 101, 47, 104, 112, 104, 112, 47, 116, 101, 115, 116, 47, 113, 117, 105, 99, 107, 47, 114, 101, 100, 101, 99, 108, 97, 114, 101, 100, 95, 99, 108, 97, 115, 115, 46, 49, 46, 105, 110, 99])
```
The `includes` of a symbol_ref may be any type of `IncludePath`, but is printed as its `Absolute` version, so it is not/doesn't seem possible to retrieve the underlying `IncludePath` of the original `HackCUnit`. So we decided that an `Absolute` path is equal to a `SearchPathRelative` as long as the `Absolute` path ends in that `SearchPathRelative`.

Reviewed By: aorenste

Differential Revision: D37761660

fbshipit-source-id: f83c11763b520dd27dc9d3782eb6de40b5de3cc2