refactor AffectedDeps prior to adding telemetry
Summary:
Functions in Shallow_class_fanout usually take an AffectedDeps.t as input and add more deps to it. They're called `add_xxx_fanout` because they add to the input dep sets. In this diff, we change the names to `get_xxx_fanout` and the signature so that the functions don't get an AffectedDeps.t as input. Instead, the unioning of all dep sets is done at the end.

This is so that we can log separately each fanout subset and what class diff caused them.

Reviewed By: bobrenjc93

Differential Revision: D36130732

fbshipit-source-id: e99dcf0201ca9d44a75b612119ae5e6a3e556685