Add ssa pass
Summary:
The SSA pass takes in code that uses Special::Var variables and converts it into SSA phi node form (using JmpArgs when necessary).

Note that it does not convert Hhbc LocalIds to SSA variables - those represent user variables and need to be preserved (or at least writes to them need to be preserved).

The SSA code itself is based on Hack Native's SSA pass.

Reviewed By: edwinsmith

Differential Revision: D37014932

fbshipit-source-id: 5513fe3706d7eee85b3681ba665f870f26a85f56