Consolidate access to direct decl parser in Decl_provider
Summary:
All the usages of the direct decl parser in `Decl_provider` follow a pattern of parsing a file, then running `List.find_map` on the result to find a specific definition.

This diff consolidates this logic in `Decl_provider.find_in_direct_decl_parse` and adds a flag determining whether or not the direct decl parser should cache the parsed decls. Note that the flag is currently always true, I'm passing `false` when making changes in subsequent diffs.

Reviewed By: andrewjkennedy

Differential Revision: D37344485

fbshipit-source-id: b5b9af3016b8716ecb4c2a877e8cd26c4ab22094