Pessimised_shared_memory backend provider
Summary:
This adds a new variant to `Provider_backend.t`, named `Pessimised_shared_memory`. It pessimises definitions not already found in the decl heap on-the-fly. To this end, it carries around a record (`Provider_backend.pessimisation_info`) of pessimisation functions, which are invoked on all definitions read from disk.

Note that in general, `Pessimised_shared_memory` behaves like `Shared_memory`. I have not attempted to fully consolidate the shared logic between both cases in many functions, to make it easier to remove  `Pessimised_shared_memory` eventually.

Reviewed By: andrewjkennedy

Differential Revision: D37344794

fbshipit-source-id: d3fec9704bb6ac51e66cf8f0ac88cdb7a8c83342