log single class fanouts to hh_fanout scuba table
Summary:
When single class fanouts are greater than 100,000 deps, we log them to hh_fanout_analysis table.
In addition to the fine-grained class diff, we also compute and log a diff "category". Class diffs would typically contains names, e.g. which specific interface was added as parent, whereas the category wouldn't, and would just specify that some interfaces were added. This is to be able to study which change categories are the most prevalent in production.

Reviewed By: bobrenjc93

Differential Revision: D36132369

fbshipit-source-id: 30c973d3f6b73afa2ad3dbc884c5a37441608f85