Use specialization to debug print Str nicer
Summary:
When debugging we often debug-print Slice but when that Slice represents a Str (which is an alias for Slice<u8>) it prints really ugly:
```
Slice[95,95,69,110,116,114,121,80,111,105,110,116,]
```
This change uses min_specialization to print `Slice` nicer when `T` is `u8`.  Unfortunately min_specialization is unstable so it has to be turned with the `UNSTABLE_DEBUG_SLICE` flag.

Differential Revision: D37822107

fbshipit-source-id: dbf9c20d944c1e4aa3ab124a856b16217b34c028