Implement canon name lookups from naming DB
Summary:
Implement lookup of type and function names by case-insensitive hash (i.e., the hash of the lowercased name).

The OCaml implementation finds canon names in a somewhat roundabout way: it uses the naming table to look up the name of the file which defines the type or function, then parses that file, searches the parse tree for a case-insensitive match, and returns the casing from the definition it found in the parse tree.

We could do this in rupro too, but I don't think it's necessary. We can get a result directly from the database: while the naming_symbols table (which contains the case-insensitive hash) doesn't contain the symbol name as a string, the naming_file_info table _does_ have strings for all the symbols in the file (separated by pipe characters).

Reviewed By: shayne-fletcher

Differential Revision: D37775377

fbshipit-source-id: 295f083f005762be755b4bff54a85ef5df78bccb