Implement OCaml serialize/deserialize for Rust_provider_backend.t
Summary:
The selected ProviderBackend is included in the server environment and included in most MultiWorker job closures, so when executing a MultiWorker job, we serialize the ProviderBackend in the master process and deserialize it in the workers. This currently leads to a crash when ProviderBackend.Rust_provider_backend is selected, since a Rust_provider_backend.t is not serializable. The Shared_memory backend solves this problem by storing everything in globals and using Provider_backend.Shared_memory as a tag (a zero-argument variant; easily serializable) indicating that those globals should be accessed.

Unfortunately, it's not clear how we could follow this pattern: we'd need to initialize our globals when workers are initialized. As far as I can tell, workers don't, at initialization time, have access to server configuration that could tell us whether to use the default Shared_memory store or Rust_provider_backend (or, further, what ParserOptions to configure the Rust_provider_backend with).

Instead, enable sending the Rust_provider_backend.t in job closures by implementing the serialize/deserialize custom operations. They're implemented by serializing the inputs necessary to construct and configure an HhServerProviderBackend (a `hackrs_provider_backend::Config`). We assume that the shm connection is already attached (since it is attached at worker-init time, before any job closures are received).

There is one place we must set a global, in Decl_store. As mentioned, we can't easily set it at worker-init time. We also can't set it during deserialization of a Rust_provider_backend.t, since the OCaml/C interface does not allow allocation during deserialization.

Instead, set the global decl store lazily whenever a toplevel decl is accessed. This should ensure that it's set properly: we should generally look up a class decl before we attempt to look up any of its members, and looking up a class decl will set the global decl store. Even if we fail to set it before an attempted member lookup, the disable_member_heaps behavior should allow us to recover: we'll see no member in the member cache, then fetch the decl of the containing shallow class in order to repopulate the member cache. Fetching that decl would set the global decl store, ensuring future lookups go straight to our FFI.

Reviewed By: shayne-fletcher

Differential Revision: D37529094

fbshipit-source-id: d3c5894137b8d375bd0fd60116221e35e6ff7111