Filter what dependencies are collected in pessimisation graph
Summary:
Until now, the "fine-grained" dependency graph used to schedule code-based pessimisation collected (almost) a strict superset of the edges found in the existing dependency graph (created by `Typing_deps`) plus inter-method/function dependencies.

This diff adds some logic to ensure that only those edges are recorded that will be used in the first version of the code-based pessimisation tool. To this end
- self-loop are always ignored
- definitions from HHI files are ignored (on a best-effort basis)
- only dependencies between functions and methods are recorded

Reviewed By: andrewjkennedy

Differential Revision: D37603332

fbshipit-source-id: 80bc65d0b9252373969487ddc8f2016ad146786c