Rename "fine-grained" dependencies
Summary:
This diff renames `Typing_fine_deps` to  `Typing_pessimisation_deps` to emphasize that these dependencies are collected for the purpose of guiding the code-based pessimisation. This feels advisable now that the set of dependencies collected in this  `Typing_fine_deps` is increasinly different from the set of dependencies collected by  `Typing_deps`.

The module `Typing_pessimisation_deps` is also moved into its own library, rather than being part of the same library as `Typing_deps`.

Reviewed By: andrewjkennedy

Differential Revision: D37603505

fbshipit-source-id: a68cbf8e457d88512f7605e972efae1b9710bc85