Flag for not caching ASTs
Summary:
This diff adds the field `allow_ast_caching` to the record `Provider_backend.pessimisation_info`. The latter contains information determining the behavior of various `*_provider`s when using `Provider_backend.Pessimised_shared_memory` as the backend.

This new flag determines the behavior of `Ast_provider` when using the `Provider_backend.Pessimised_shared_memory` backend: If it is unset, no ASTs must be read from or written to any file or AST cache. This is used to implement reading pessimised ASTs from a separate repository in subsequent diffs.

Reviewed By: andrewjkennedy

Differential Revision: D37647981

fbshipit-source-id: 4f5e88bbfeea822c730dc22d97b10f650fb395a0