Adding a lint for Unconditional Recursion: checking the statements that appear before the recursive call
Summary:
This produces a lint error of unconditional recursion when there is a top level recursive call and there is no return or throw statement to break out of the recursion before the recursive call. The return or throw statement could be within multiple nested conditionals. The expression following the return/throw statement is checked to ensure that it is anything other than a recursive call. If there are mutliple base cases, then it is considered unconditional recursion if each of the return or throw statements in the base cases are recursive function calls.
The function/method may or may not have arguments.
The different types of function/method calls detected by linter are:
1) return statements
2) throw statements
3) static methods
4) instance methods
5) Await

Reviewed By: Wilfred

Differential Revision: D37754411

fbshipit-source-id: c2767d1df54350072b8663de5705d786b7aa5313