Allow like-typed default parameter expression even for non-enforceable types
Summary:
For a default parameter expression
```
<<__SupportDynamicType>>
function foo(t $x = e)
```
with declared parameter type `t`, we should allow the expression `e` to have any type `u` such that `u <: ~t`.

If `t` is enforced, then the parameter is assumed to have that type when checking the body.
If `t` is not enforced, then the parameter is assumed to have type `u|t` when checking the body, as though the default expression had been assigned conditionally to the parameter.

Reviewed By: vassilmladenov

Differential Revision: D37854960

fbshipit-source-id: bfa435b0b6f802b24606f58f9c198e1a5c6b3095