Use std::filesystem consistently
Summary:
For filesystem manipulation, we use a mix of boost::filesystem and
folly::fs. However, folly::fs could be boost::filesystem or
std::filesystem depending on which header you include, and we use
both. This means you can get strange compiler errors or linking errors
depending on which headers you include (or what order they're included
in).

Unify around std::filesystem and use that
consistently. boost::filesystem was the prototype for std::filesystem
so most things are just a drop in replacement. Some of the implicit
conversions are gone, and a few methods are renamed/removed. Fix
accordingly.

The only piece of boost::filesystem we still use is unique_path()
because there's no replacement for that in std::filesystem yet.

Reviewed By: mofarrell

Differential Revision: D37876525

fbshipit-source-id: 0e5308de61cb4a28cf0854b51a79f164560753f4