Soft exception for unimplemented features
Summary:
To distinguish between hard failures unrelated to the analysis and analysis failures, we use Shape_analysis_exn. Within the walker code, this is thrown through calls to `failwithpos`, but in two places I was using `failwith` which throws a generic exception that is not suppressed later on.

To prevent hard exceptions preemptively killing the analysis, this diff fixes the uses of `failwith` to use `Shape_analysis_exn`

Differential Revision: D37853559

fbshipit-source-id: 1bf2843210491667da277fb450011d3a43bb4a9d