Add debugging information to constraints
Summary:
It was getting difficult to debug the generated constraints. I changed the constraint collection logic so that we note the Hack source position of the constraint as well as the line in the OCaml source code where the constraint is generated.

The line numbers for the OCaml source code refer exclusively to Shape_analysis_walker even if the actual constraint is generated in a different module such as Shape_analysis_environment.

The OCaml line numbers are only displayed when `--verbosity` flag is set to 0. Notably, it is omitted hence set to 0 in tests. Otherwise, irrelevant changes to constraint generation affects the expected test output.

The ordering of constraints follows the line number of the Hack source code. This should make it easier to read test cases as well as debug problems.

The vast majority of this diff is plumbing location information through the codebase.

Differential Revision: D37715094

fbshipit-source-id: 824ca8a7b5eae66de0f08fd31e262be5a55af7e2