Fix loop folding via constraints bug
Summary:
Environment looping for loops was buggy in a very interesting way. This diff fixes it.

Consider the following fragment where the first dictionary is allocated at position `p` and the second one at `q`.
```
$d = dict['p' => 42];
while(...) {
  $d = dict['q' => true];
}
```

We were generating the constraints,
```
p -> #1 // line 1
q -> #2 // line 3
#1 \/ #2 -> #3 // line 4 on exit
#2 -> #1 // line 2-4
```
The last constraint is added because the definition of `$d` at the end of the loop needs to be related to the beginning of the loop.

However, this implies the following
```
p -> #1
q -> #1
q -> #2
```

The way we compute optional fields is if the join point has one entity in one branch but not in the other. So the solver concluded that `p` was an optional field but `q` was not (despite the fact that it is optional due to the fact that the loop body might never run)!

The problem is that we were using the same environment for at the beginning of the loop body and right before it. The solution is to simply create a redirection when we start typechecking the loop body and loop back onto that. This leads to the constraints:

```
p -> #1 // line 1
#1 -> #2 // due to refresh
q -> #3 // line 3
#1 \/ #3 -> #4 // line 4 on exit
#3 -> #2 // line 2-4
```

Then the solver concludes
```
p -> #1
q -> #3
q -> #2
```

Consequently, the solver concludes that at the join point `#4` both `p` and `q` are optional as one doesn't exist in `#1` and the other doesn't exist in `#3`.

Differential Revision: D37689958

fbshipit-source-id: 66cc8242fd87d99b5c495ff2747d60ba04a60a5a