Partially move methods returning `this` in Hack Collections to PHP
Summary:
I had originally done this to unblock typing some native functions that return `this`; I realized I don't *need* to do such, but I noticed that this was net less code and avoids relying on "all unknown types are `KindOfObject`."

This pushes the Hack Collection builtins that return `$this` slightly into userland. All of them are replaced with wrappers that call a native function returning `void` (doing any mutation needed) and then immediately returning `$this`.

Reviewed By: ricklavoie

Differential Revision: D37657963

fbshipit-source-id: 6ddca5d09f6892bf968955f81b2140f720fd32e2