Add support for idx in shape-like-dict analysis
Summary:
This diff is a first step towards supporting the special indexing API `idx` for dicts (see https://docs.hhvm.com/hack/reference/function/HH.idx/) in the shape-like-dict analysis.

For now, we treat `idx` expressions with and without default value in the same way.

Essentially, we follow the case for `A.Array_get`, after extracting the right data in the `A.call` case.

To test the error message on `line 200`, I added a test function that makes use of an `idx` call with only one argument. Since this lead to an immediate type error, thus not triggering the custom error message, I put the function in comments. Remarks on how to test `line 200` are appreciated.

Also, I am surprised that `id.php.exp` lists the same shape at every inspection. I would have expected that the inspection on line 4 leads to an empty shape, the inspection on line 6 to a shape with one key, etc.

Reviewed By: madgen

Differential Revision: D37924482

fbshipit-source-id: 4baa3d09dfda1a53d5fc6c1feb1d2745573b5de8