third-party/rust: update md-5, sha-1, sha2, digest to 0.10; hmac 0.12
Summary:
It's time.

Lots of updates, mostly:
- `input()` -> `update()`
- `result()` -> `finalize()`

Small update to `Hmac`:
- `new_varkey` -> `new_from_slice`
- drop `HmacNew`

We have a couple of versions of curve25519-dalek which have tight couplings to
specific versions of `digest`, so we need to keep corresponding old sha2 crates
around.

Also includes some file deletions which were apparently missed in a previous
import.

Reviewed By: dtolnay

Differential Revision: D37948669

fbshipit-source-id: 22bc87c627776e1fdd0af321a576bcfb58409e2f