Add SDT annotation to some hhi functions
Summary:
The functions `array_mark_legacy` are generic functions that should not be pessimised, as they can be used in intializer expressions for constants where we don't want to add like types. For example, we want to permit
```
class C {
  const vec<int> VI = HH\array_mark_legacy(vec[1]);
}
```
but this would be rejected if the result of `array_mark_legacy` could be dynamic.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D37930228

fbshipit-source-id: 10984e9de6a7433108099bd301f83412cc81fc07