Improve mapping in of data into core file
Summary:
This makes it so we can map in partial jit.prof files, which are actually more
common than the complete ones in many cases.  It adds a utility for pulling the
file back out of the core.  While here I add the stacktrace file into the core
for ease of access in annoying circumstances.

Reviewed By: binliu19

Differential Revision: D37975051

fbshipit-source-id: 92b8205331971707949c8cf4dba59fdc11826fd5