Lambda parameter should default to dynamic under dynamic check
Summary: When checking a lambda in an SDT function under dynamic assumptions, we attempt to "infer" a type for the lambda parameter but often have no information other than a `<: supportdyn<mixed>` bound. In these circumstances, rather than raise an error ("Was expecting an object but type is unknown") we should assume `dynamic` for the type of the receiver.

Differential Revision: D37975578

fbshipit-source-id: 99d134fe6bbf3ea2ebb62e99f2d2bc91cd8141cd