Iterate simplify to remove all unreachable code
Summary:
Simplify is the pass that actually makes instructions dealing in Bottoms
Unreachable and prunes them from the unit.  In pruning blocks, it may improve
other types leading to more Bottoms that can be pruned.  This is not just dead
code removal, but a correctness issue.  Some instructions' codegen is unable to
handle bottom inputs and must be correctly removed prior to irlower.

Reviewed By: ricklavoie

Differential Revision: D38003677

fbshipit-source-id: 9ae0aaaca197671ff4616b44d7548e4fa0bfbc79