prevent infinite recursion when accessing non-existent static methods
Summary:
if class `C` uses a trait that `require class C`, then the flattened decls include `C` in the list of `all_ancestor_req_class_requirements` for class `C`.  If we access a non-existent static method `foo` on class `C`, the search algorithm was entering an infinite recursion:
- static method `foo` cannot be resolved using decl `smethods`, but...
- since `C` has an upper bound on `this` from class-level where clauses or `require class` constraints search for `foo` in all the upper bounds, but...
- `C` itself is an upper bound via flattened `require class` constraints, so restart the search in `C`...

A similar issue happens with class-level where clauses, if the programmer specifies an explicit `where this as C` bound for class C.

This diff ensures that upper bounds equivalent (after localization to take into account type aliases) to the current class are removed from the list of upper bounds, preventing the infinite recursion.

Reviewed By: vassilmladenov

Differential Revision: D37600573

fbshipit-source-id: a92d5fd092d0745ca9895e312234ac9587bef2dd