Support HH\Vectors
Summary:
In this diff, I added support for mutable Vectors in the walker. Since Vectors are mutable and copy-by-reference, we should not create a new entity to represent them after an array assignment, since the original Vector is still being used. Hence, the new test `vector2.php` below should report two upcasts, since both variables in the file point to the same physical vector containing `f<>`.

I also added comments clarifying that we are handling copy-on-write for vecs during assignment.

Next steps is to only report upcasts for function pointers that are later called. If an upcasted function pointer is not used, there is no need to report it, since a programmer can refactor the original function without causing issues due to the upcasted function pointer.

Reviewed By: madgen

Differential Revision: D37971136

fbshipit-source-id: a3c899d7409ed51cfa7cff2f74aef364b517f64d