Add function calls to walker and fix tests to include function calls (1/2)
Summary:
The walker now recurses into function calls and their arguments. I also changed the simplify_constraints tests to include calls on the upcasted function pointers.

I decided against changing the dump_constraints tests to also include calls, since the main point of those tests are to make sure that the walker is handling constraints correctly in those situations. On the other hand, the simplify_constraint tests are meant to make sure that upcasts are reported correctly. If I don't make this change on the tests, the simplify_constraint tests will report no upcasts after my next few diffs.

A future diff will add a new constraint to track function calls, and I will create a new dump_constraint test for that constraint.

Context:
Since this is ultimately a user-facing tool to help programmers refactor their code safely, our goal is to report upcasts dynamically. If an upcast occurs and the upcasted function pointer is never used, there is no need to report the upcast, since it does nothing.

The next diff fixes a test, but the diff after that will implement a new constraint to enable us to only report upcasts that are called.

Reviewed By: madgen

Differential Revision: D37983085

fbshipit-source-id: ca7b0c18403cff675048e5bf3d6558356363d988