Check for calls before reporting upcasts (2/2)
Summary:
The solver and walker now only reports upcasts to function pointers that are later called. This is because someone refactoring their code doesn't need to worry about upcasted function pointers that are never used.

To do this, I added a new constraint `Called` that detects whether an entity has been called. In the solver, we check that an entity that was `Called` is the superset of an upcast, meaning that a function call occurred after an upcast.

Differential Revision: D38015167

fbshipit-source-id: ffd54bc651bb653b1b6be33dbcc5872ab38d343c