Always error on nested Expression Trees
Summary:
Currently, we have checks that are intended to prevent nested expression trees because that functionality isn't anticipated / quite work.

We have a check in the parser and we have a check in the desugarer. The parser check is mostly to resolve an ambiguity between nested expression trees and constants within expression trees.

The desugarer check is not complete because we weren't recursing into the splices. This change adds a recursive pass into the splices to check for nested expression trees in splices.

```
ExampleDSL`() ==> {
  $x = ${ ExampleDSL`1` };
}`
```

Reviewed By: Wilfred

Differential Revision: D37903931

fbshipit-source-id: 7f1d9a83cc508cf5334fa6dbb637fbe0df6f1263